--- main/java/board/list/BoardListServlet.java
@@ -1,28 +1,0 @@
-package board.list;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.Date;
-import java.util.List;
-
-import javax.servlet.ServletException;
-//import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-//@WebServlet("/boardlist")
-public class BoardListServlet extends HttpServlet {
-
-	private static final long serialVersionUID = 1L;
-	@Override
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) 
-	throws ServletException, IOException {
-		response.setContentType("text/html;charset=utf-8");
-		PrintWriter out = response.getWriter();
-		BoardListDAO dao = new BoardListDAO();
-		List<BoardListVO> list = dao.listBoard(1);
-	
-		
-	}
-}
