--- main/webapp/WEB-INF/web.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="WebApp_ID" version="3.0">
-  <display-name>pro07</display-name>
-  <welcome-file-list>
-    <welcome-file>index.html</welcome-file>
-    <welcome-file>index.htm</welcome-file>
-    <welcome-file>index.jsp</welcome-file>
-    <welcome-file>default.html</welcome-file>
-    <welcome-file>default.htm</welcome-file>
-    <welcome-file>default.jsp</welcome-file>
-  </welcome-file-list>
-
-	<!-- 게시글 작성 -->
-	<servlet>
-		<servlet-name>boardregister</servlet-name>
-		<servlet-class>board.register.BoardRegisterServlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>boardregister</servlet-name>
-		<url-pattern>/boardregister</url-pattern>
-	</servlet-mapping>
-	
-	
-
-</web-app>
