+++ src/main/webapp/WEB-INF/classes/board/detail/BoardDetailVO.class
